Fix some more compile errors

+++ terra-offheap/src/main/java/com/ritualsoftheold/terra/offheap/chunk/compress/EmptyChunkFormat.java
@@ -22,7 +22,7 @@ public class EmptyChunkFormat implements ChunkFormat {
     }
 
     @Override
-    public int getChunkType() {
+    public byte getChunkType() {
         return ChunkType.EMPTY;
     }
