+++ terra-offheap/src/main/java/com/ritualsoftheold/terra/offheap/chunk/compress/Palette16ChunkFormat.java
@@ -163,7 +163,7 @@ public class Palette16ChunkFormat implements ChunkFormat {
     }
 
     @Override
-    public int getChunkType() {
+    public byte getChunkType() {
         return ChunkType.PALETTE16;
     }
